Remove cyclical dependency from product type
