+++ support-frontend/assets/components/checkoutNudge/checkoutNudgeContainer.tsx
@@ -6,13 +6,11 @@ import { detect, glyph } from 'helpers/internationalisation/currency';
 import { setProductType } from 'helpers/redux/checkout/product/actions';
 import {
 	getContributionType,
-	getSelectedAmount,
-} from 'helpers/redux/checkout/product/selectors/productType';
-import {
 	getMaximumContributionAmount,
 	getMinimumContributionAmount,
-	isUserInAbVariant,
-} from 'helpers/redux/commonState/selectors';
+	getSelectedAmount,
+} from 'helpers/redux/checkout/product/selectors/productType';
+import { isUserInAbVariant } from 'helpers/redux/commonState/selectors';
 import {
 	useContributionsDispatch,
 	useContributionsSelector,
