+++ support-frontend/assets/helpers/redux/commonState/selectors.ts
@@ -1,7 +1,4 @@
 import type { ContributionType } from 'helpers/contributions';
-import { config } from 'helpers/contributions';
-// eslint-disable-next-line import/no-cycle -- these are quite tricky to unpick so we should come back to this
-import { getContributionType } from '../checkout/product/selectors/productType';
 import type { ContributionsState } from '../contributionsStore';
 
 export function getDefaultContributionType(
@@ -9,30 +6,6 @@ export function getDefaultContributionType(
 ): ContributionType {
 	const { defaultContributionType } = state.common.amounts;
 	return defaultContributionType;
-}
-
-export function getMinimumContributionAmount(
-	contributionType?: ContributionType,
-) {
-	return (state: ContributionsState): number => {
-		const { countryGroupId } = state.common.internationalisation;
-		const { min } =
-			config[countryGroupId][contributionType ?? getContributionType(state)];
-
-		return min;
-	};
-}
-
-export function getMaximumContributionAmount(
-	contributionType?: ContributionType,
-) {
-	return (state: ContributionsState): number => {
-		const { countryGroupId } = state.common.internationalisation;
-		const { max } =
-			config[countryGroupId][contributionType ?? getContributionType(state)];
-
-		return max;
-	};
 }
 
 export function isUserInAbVariant(abTestName: string, variantName: string) {
